requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2

# Gmail / Calendar
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Todoist
todoist-api-python>=2.1.0

# WhatsApp via Twilio (swap for your provider of choice)
twilio>=9.0.0
